ML classification models testing
This code contains the test I made using classification models to assert Creativity and Innovation Techniques, using multioutput models with predict probability (3 highest CITs probabilities are outputted if reached a certain threshold). The best model was chosen based on a selfmade metric sistem using accuracy, sum of square means of probabilities, and number of empty outputs (which did not reach the minimal threshold. Parameters of the best model were optimized and plotted